Reorganize historian.py by removing History and types to other files. Also optimize imports in all src/ files
