+++ src/quest/history.py
@@ -1,0 +1,12 @@
+from typing import Protocol, Reversible
+from src.quest.types import EventRecord
+
+
+class History(Protocol, Reversible):
+    def append(self, item: EventRecord): ...
+
+    def remove(self, item: EventRecord): ...
+
+    def __iter__(self): ...
+
+    def __reversed__(self): ...
